added a nextjs front end to connect to the pocketbase instance.. lots of new files. login, signup, and signout all work.. so does updating the user's name ( so cru ) havent tried delete yet.
Some decisions need to be made around structure. I struggled trying to get pocketbase to play nice on the server AND client.. and I think I need to rely on the server components more for the data layer instead of shoe-horning it into the client side.

shad-cn is nice. I like that.